api/requirements.txt
flask==1.1.4
werkzeug==1.0.1
flask-cors==4.0.0
gunicorn==21.2.0
flask-pwa==0.1.0
Pillow==10.1.0
markupsafe==2.0.1
jinja2==2.11.3
tensorflow==2.15.0
tensorflow-hub==0.15.0
numpy==1.24.3
pandas==2.1.3